refactor: more helper objects for dedup
