+++ src/cli/repl/commands/repl-cfg.ts
@@ -6,6 +6,7 @@ import type { NormalizedAst } from '../../../r-bridge/lang-4.x/ast/model/process
 import { type CfgSimplificationPassName, DefaultCfgSimplificationOrder } from '../../../control-flow/cfg-simplification';
 import type { ReadonlyFlowrAnalysisProvider } from '../../../project/flowr-analyzer';
 import { handleString } from '../core';
+import { ReplClipboard } from './repl-clipboard';
 
 function formatInfo(out: ReplOutput, type: string): string {
 	return out.formatter.format(`Copied ${type} to clipboard.`, { color: Colors.White, effect: ColorEffect.Foreground, style: FontStyles.Italic });
@@ -15,14 +16,7 @@ async function produceAndPrintCfg(analyzer: ReadonlyFlowrAnalysisProvider, outpu
 	const cfg = await analyzer.controlflow([...DefaultCfgSimplificationOrder, ...simplifications]);
 	const normalizedAst = await analyzer.normalize();
 	const mermaid = cfgConverter(cfg, normalizedAst);
-	output.stdout(mermaid);
-	if(output.allowClipboard !== false) {
-		try {
-			const clipboard = await import('clipboardy');
-			clipboard.default.writeSync(mermaid);
-			output.stdout(formatInfo(output, 'mermaid code'));
-		} catch{ /* do nothing this is a service thing */ }
-	}
+	await ReplClipboard.print(output, mermaid, formatInfo(output, 'mermaid code'));
 }
 
 export const controlflowCommand: ReplCodeCommand = {
