+++ src/control-flow/diff-cfg.ts
@@ -1,6 +1,5 @@
-import { jsonReplacer } from '../util/json';
 import type { NodeId } from '../r-bridge/lang-4.x/ast/model/processing/node-id';
-import { type GraphDiffContext, type NamedGraph, initDiffContext, GraphDifferenceReport } from '../util/diff-graph';
+import { type GraphDiffContext, type NamedGraph, initDiffContext, GraphDifferenceReport, GraphDiff } from '../util/diff-graph';
 import { type GenericDiffConfiguration, setDifference } from '../util/diff';
 import { CfgEdge, CfgVertex, type ControlFlowGraph } from './control-flow-graph';
 import { arrayEqual } from '../util/collections/arrays';
@@ -25,7 +24,7 @@ export function diffOfControlFlowGraphs(left: NamedGraph<ControlFlowGraph>, righ
 function diffDataflowGraphs(ctx: GraphDiffContext<ControlFlowGraph>): void {
 	diffRootVertices(ctx);
 	diffVertices(ctx);
-	diffOutgoingEdges(ctx);
+	GraphDiff.outgoingEdges(ctx, diffEdges);
 }
 
 function diffRootVertices(ctx: GraphDiffContext<ControlFlowGraph>): void {
@@ -113,41 +112,6 @@ function diffVertices(ctx: GraphDiffContext<ControlFlowGraph>): void {
 	}
 }
 
-function diffOutgoingEdges(ctx: GraphDiffContext<ControlFlowGraph>): void {
-	const lEdges = new Map([...ctx.left.edges()]);
-	const rEdges = new Map([...ctx.right.edges()]);
-
-	if(lEdges.size < rEdges.size && !ctx.config.leftIsSubgraph || lEdges.size > rEdges.size && !ctx.config.rightIsSubgraph) {
-		ctx.report.addComment(`Detected different number of edges! ${ctx.leftname} has ${lEdges.size} (${JSON.stringify(lEdges, jsonReplacer)}). ${ctx.rightname} has ${rEdges.size} ${JSON.stringify(rEdges, jsonReplacer)}`);
-	}
-
-	for(const [id, edge] of lEdges) {
-		/* This has nothing to do with the subset relation as we verify this in the same graph.
-         * Yet we still do the check as a subgraph may not have to have all source vertices for edges.
-         */
-		if(!ctx.left.hasVertex(id)) {
-			if(!ctx.config.leftIsSubgraph) {
-				ctx.report.addComment(`The source ${id} of edges ${JSON.stringify(edge, jsonReplacer)} is not present in ${ctx.leftname}. This means that the graph contains an edge but not the corresponding vertex.`);
-				continue;
-			}
-		}
-		diffEdges(ctx, id, edge, rEdges.get(id));
-	}
-	// just to make it both ways in case the length differs
-	for(const [id, edge] of rEdges) {
-		if(!ctx.right.hasVertex(id)) {
-			if(!ctx.config.rightIsSubgraph) {
-				ctx.report.addComment(`The source ${id} of edges ${JSON.stringify(edge, jsonReplacer)} is not present in ${ctx.rightname}. This means that the graph contains an edge but not the corresponding vertex.`);
-				continue;
-			}
-		}
-		if(!ctx.config.leftIsSubgraph && !lEdges.has(id)) {
-			diffEdges(ctx, id, undefined, edge);
-		}
-		/* otherwise, we already cover the edge above */
-	}
-}
-
 function diffEdge(edge: CfgEdge, otherEdge: CfgEdge, ctx: GraphDiffContext<ControlFlowGraph>, id: NodeId, target: NodeId) {
 	const el = CfgEdge.getType(edge);
 	const ol = CfgEdge.getType(otherEdge);
@@ -176,40 +140,5 @@ function diffEdge(edge: CfgEdge, otherEdge: CfgEdge, ctx: GraphDiffContext<Contr
 }
 
 function diffEdges(ctx: GraphDiffContext<ControlFlowGraph>, id: NodeId, lEdges: ReadonlyMap<NodeId, CfgEdge> | undefined, rEdges: ReadonlyMap<NodeId, CfgEdge> | undefined): void {
-	if(lEdges === undefined || rEdges === undefined) {
-		if(
-			(lEdges === undefined && !ctx.config.leftIsSubgraph)
-            || (rEdges === undefined && !ctx.config.rightIsSubgraph)
-		) {
-			ctx.report.addComment(
-				`Vertex ${id} has undefined outgoing edges. ${ctx.leftname}: ${JSON.stringify(lEdges, jsonReplacer)} vs ${ctx.rightname}: ${JSON.stringify(rEdges, jsonReplacer)}`,
-				{ tag: 'vertex', id }
-			);
-		}
-		return;
-	}
-
-	if(
-		lEdges.size < rEdges.size && !ctx.config.leftIsSubgraph
-        || lEdges.size > rEdges.size && !ctx.config.rightIsSubgraph
-	) {
-		ctx.report.addComment(
-			`Vertex ${id} differs in number of outgoing edges. ${ctx.leftname}: [${[...lEdges.keys()].join(',')}] vs ${ctx.rightname}: [${[...rEdges.keys()].join(',')}] `,
-			{ tag: 'vertex', id }
-		);
-	}
-	// order independent compare
-	for(const [target, edge] of lEdges) {
-		const otherEdge = rEdges.get(target);
-		if(otherEdge === undefined) {
-			if(!ctx.config.rightIsSubgraph) {
-				ctx.report.addComment(
-					`Target of ${id}->${target} in ${ctx.leftname} is not present in ${ctx.rightname}`,
-					{ tag: 'edge', from: id, to: target }
-				);
-			}
-			continue;
-		}
-		diffEdge(edge, otherEdge, ctx, id, target);
-	}
+	GraphDiff.edges(ctx, id, lEdges, rEdges, diffEdge);
 }
