+++ src/core/steps/all/core/11-normalize-tree-sitter.ts
@@ -1,14 +1,8 @@
-import { internalPrinter, StepOutputFormat } from '../../../print/print';
-import {
-	normalizedAstToJson,
-	normalizedAstToQuads,
-	printNormalizedAstToMermaid,
-	printNormalizedAstToMermaidUrl
-} from '../../../print/normalize-printer';
 import { type IPipelineStep, PipelineStepStage } from '../../pipeline-step';
 import type { DeepReadonly } from 'ts-essentials';
 import { normalizeTreeSitter } from '../../../../r-bridge/lang-4.x/ast/parser/json/parser';
 import type { NormalizeRequiredInput } from './10-normalize';
+import { NormalizePrinters } from './10-normalize';
 import type { ParseStepOutput } from '../../../../r-bridge/parser';
 import type { Tree } from 'web-tree-sitter';
 import type { FlowrAnalyzerContext } from '../../../../project/context/flowr-analyzer-context';
@@ -23,13 +17,7 @@ export const NORMALIZE_TREE_SITTER = {
 	description:       'Normalize the AST to flowR\'s AST',
 	processor,
 	executed:          PipelineStepStage.OncePerFile,
-	printer:           {
-		[StepOutputFormat.Internal]:   internalPrinter,
-		[StepOutputFormat.Json]:       normalizedAstToJson,
-		[StepOutputFormat.RdfQuads]:   normalizedAstToQuads,
-		[StepOutputFormat.Mermaid]:    printNormalizedAstToMermaid,
-		[StepOutputFormat.MermaidUrl]: printNormalizedAstToMermaidUrl
-	},
-	dependencies:  [ 'parse' ],
-	requiredInput: undefined as unknown as NormalizeRequiredInput
+	printer:           NormalizePrinters,
+	dependencies:      [ 'parse' ],
+	requiredInput:     undefined as unknown as NormalizeRequiredInput
 } as const satisfies DeepReadonly<IPipelineStep<'normalize', typeof processor>>;
