+++ src/dataflow/graph/diff-dataflow-graph.ts
@@ -11,6 +11,7 @@ import { diffEnvironmentInformation, diffIdentifierReferences } from '../environ
 import { EmptyArgument } from '../../r-bridge/lang-4.x/ast/model/nodes/r-function-call';
 import { diffControlDependencies } from '../info';
 import type { GraphDifferenceReport, GraphDiffContext } from '../../util/diff-graph';
+import { GraphDiff } from '../../util/diff-graph';
 import type { HookInformation } from '../hooks';
 
 
@@ -21,42 +22,7 @@ import type { HookInformation } from '../hooks';
 export function diffDataflowGraph(ctx: GraphDiffContext): void {
 	diffRootVertices(ctx);
 	diffVertices(ctx);
-	diffOutgoingEdges(ctx);
-}
-
-function diffOutgoingEdges(ctx: GraphDiffContext): void {
-	const lEdges = new Map(ctx.left.edges());
-	const rEdges = new Map(ctx.right.edges());
-
-	if(lEdges.size < rEdges.size && !ctx.config.leftIsSubgraph || lEdges.size > rEdges.size && !ctx.config.rightIsSubgraph) {
-		ctx.report.addComment(`Detected different number of edges! ${ctx.leftname} has ${lEdges.size} (${JSON.stringify(lEdges, jsonReplacer)}). ${ctx.rightname} has ${rEdges.size} ${JSON.stringify(rEdges, jsonReplacer)}`);
-	}
-
-	for(const [id, edge] of lEdges) {
-		/* This has nothing to do with the subset relation as we verify this in the same graph.
-		 * Yet we still do the check as a subgraph may not have to have all source vertices for edges.
-		 */
-		if(!ctx.left.hasVertex(id)) {
-			if(!ctx.config.leftIsSubgraph) {
-				ctx.report.addComment(`The source ${id} of edges ${JSON.stringify(edge, jsonReplacer)} is not present in ${ctx.leftname}. This means that the graph contains an edge but not the corresponding vertex.`);
-				continue;
-			}
-		}
-		diffEdges(ctx, id, edge, rEdges.get(id));
-	}
-	// just to make it both ways in case the length differs
-	for(const [id, edge] of rEdges) {
-		if(!ctx.right.hasVertex(id)) {
-			if(!ctx.config.rightIsSubgraph) {
-				ctx.report.addComment(`The source ${id} of edges ${JSON.stringify(edge, jsonReplacer)} is not present in ${ctx.rightname}. This means that the graph contains an edge but not the corresponding vertex.`);
-				continue;
-			}
-		}
-		if(!ctx.config.leftIsSubgraph && !lEdges.has(id)) {
-			diffEdges(ctx, id, undefined, edge);
-		}
-		/* otherwise, we already cover the edge above */
-	}
+	GraphDiff.outgoingEdges(ctx, diffEdges);
 }
 
 function diffRootVertices(ctx: GraphDiffContext): void {
@@ -357,40 +323,5 @@ function diffEdge(edge: DfEdge, otherEdge: DfEdge, ctx: GraphDiffContext, id: No
  * Compares two sets of outgoing edges and reports differences.
  */
 export function diffEdges(ctx: GraphDiffContext, id: NodeId, lEdges: OutgoingEdges | undefined, rEdges: OutgoingEdges | undefined): void {
-	if(lEdges === undefined || rEdges === undefined) {
-		if(
-			(lEdges === undefined && !ctx.config.leftIsSubgraph)
-			|| (rEdges === undefined && !ctx.config.rightIsSubgraph)
-		) {
-			ctx.report.addComment(
-				`Vertex ${id} has undefined outgoing edges. ${ctx.leftname}: ${JSON.stringify(lEdges, jsonReplacer)} vs ${ctx.rightname}: ${JSON.stringify(rEdges, jsonReplacer)}`,
-				{ tag: 'vertex', id }
-			);
-		}
-		return;
-	}
-
-	if(
-		lEdges.size < rEdges.size && !ctx.config.leftIsSubgraph
-		|| lEdges.size > rEdges.size && !ctx.config.rightIsSubgraph
-	) {
-		ctx.report.addComment(
-			`Vertex ${id} differs in number of outgoing edges. ${ctx.leftname}: [${[...lEdges.keys()].join(',')}] vs ${ctx.rightname}: [${[...rEdges.keys()].join(',')}] `,
-			{ tag: 'vertex', id }
-		);
-	}
-	// order independent compare
-	for(const [target, edge] of lEdges) {
-		const otherEdge = rEdges.get(target);
-		if(otherEdge === undefined) {
-			if(!ctx.config.rightIsSubgraph) {
-				ctx.report.addComment(
-					`Target of ${id}->${target} in ${ctx.leftname} is not present in ${ctx.rightname}`,
-					{ tag: 'edge', from: id, to: target }
-				);
-			}
-			continue;
-		}
-		diffEdge(edge, otherEdge, ctx, id, target);
-	}
+	GraphDiff.edges(ctx, id, lEdges, rEdges, diffEdge);
 }
