+++ src/project/plugins/file-plugins/flowr-analyzer-description-file-plugin.ts
@@ -11,6 +11,22 @@ export const descriptionFileLog = log.getSubLogger({ name: 'flowr-analyzer-loadi
 
 const DescriptionFilePattern = /^DESCRIPTION(\.(txt|in))?$/i;
 
+/** Access to the `DESCRIPTION` file of the analyzed project. */
+export const DescriptionFile = {
+	name: 'DescriptionFile',
+	/** The project's only `DESCRIPTION` file, `undefined` if there is none; `missing` states what cannot be done then. */
+	single(this: void, ctx: FlowrAnalyzerContext, missing: string) {
+		const descFiles = ctx.files.getFilesByRole(FileRole.Description);
+		if(descFiles.length === 0) {
+			descriptionFileLog.debug(missing);
+			return undefined;
+		} else if(descFiles.length > 1) {
+			descriptionFileLog.warn(`Found ${descFiles.length} description files, expected exactly one.`);
+		}
+		return descFiles[0];
+	}
+} as const;
+
 /**
  * This plugin provides support for R `DESCRIPTION` files.
  */
