+++ src/project/plugins/package-version-plugins/flowr-analyzer-meta-description-file-plugin.ts
@@ -1,10 +1,7 @@
 import { FlowrAnalyzerPackageVersionsPlugin } from './flowr-analyzer-package-versions-plugin';
-import {
-	descriptionFileLog
-} from '../file-plugins/flowr-analyzer-description-file-plugin';
+import { DescriptionFile } from '../file-plugins/flowr-analyzer-description-file-plugin';
 import { SemVer } from 'semver';
 import type { FlowrAnalyzerContext } from '../../context/flowr-analyzer-context';
-import { FileRole } from '../../context/flowr-file';
 import { MetaPriority } from '../../context/flowr-analyzer-meta-context';
 
 /**
@@ -16,16 +13,10 @@ export class FlowrAnalyzerMetaDescriptionFilePlugin extends FlowrAnalyzerPackage
 	public readonly version = new SemVer('0.1.0');
 
 	process(ctx: FlowrAnalyzerContext): void {
-		const descFiles = ctx.files.getFilesByRole(FileRole.Description);
-		if(descFiles.length === 0) {
-			descriptionFileLog.debug('No description file found, cannot extract package versions.');
+		const deps = DescriptionFile.single(ctx, 'No description file found, cannot extract package versions.');
+		if(deps === undefined) {
 			return;
-		} else if(descFiles.length > 1) {
-			descriptionFileLog.warn(`Found ${descFiles.length} description files, expected exactly one.`);
 		}
-
-		/** this will do the caching etc. for me */
-		const deps = descFiles[0];
 
 		const pkg = deps.packageName();
 		ctx.meta.contribute({
