+++ src/project/plugins/package-version-plugins/flowr-analyzer-package-versions-description-file-plugin.ts
@@ -1,11 +1,8 @@
 import { FlowrAnalyzerPackageVersionsPlugin } from './flowr-analyzer-package-versions-plugin';
-import {
-	descriptionFileLog
-} from '../file-plugins/flowr-analyzer-description-file-plugin';
+import { DescriptionFile } from '../file-plugins/flowr-analyzer-description-file-plugin';
 import { SemVer } from 'semver';
 import type { Package } from './package';
 import type { FlowrAnalyzerContext } from '../../context/flowr-analyzer-context';
-import { FileRole } from '../../context/flowr-file';
 
 
 /**
@@ -18,16 +15,10 @@ export class FlowrAnalyzerPackageVersionsDescriptionFilePlugin extends FlowrAnal
 	public readonly version = new SemVer('0.1.0');
 
 	process(ctx: FlowrAnalyzerContext): void {
-		const descFiles = ctx.files.getFilesByRole(FileRole.Description);
-		if(descFiles.length === 0) {
-			descriptionFileLog.debug('No description file found, cannot extract package versions.');
+		const deps = DescriptionFile.single(ctx, 'No description file found, cannot extract package versions.');
+		if(deps === undefined) {
 			return;
-		} else if(descFiles.length > 1) {
-			descriptionFileLog.warn(`Found ${descFiles.length} description files, expected exactly one.`);
 		}
-
-		/** this will do the caching etc. for me */
-		const deps = descFiles[0];
 
 		this.retrieveVersionsFromField(ctx, deps.depends() ?? []);
 		this.retrieveVersionsFromField(ctx, deps.imports() ?? []);
